requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose
stripe
requests
